refactor : 이벤트 생성 폼 리팩토링
