+++ src/hooks/useAddressSearch.ts
@@ -1,0 +1,25 @@
+import { useEffect, useState } from "react";
+
+export function useAddressSearch() {
+  const [isOpen, setIsOpen] = useState(false);
+  const [isScriptLoaded, setIsScriptLoaded] = useState(false);
+
+  useEffect(() => {
+    const script = document.createElement("script");
+    script.src =
+      "//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js";
+    script.async = true;
+    script.onload = () => setIsScriptLoaded(true);
+    document.head.appendChild(script);
+
+    return () => {
+      document.head.removeChild(script);
+    };
+  }, []);
+
+  return {
+    isOpen,
+    setIsOpen,
+    isScriptLoaded,
+  };
+}
